Change names correctly, add comments
